requirements.txt
requests>=2.25.0
pyyaml>=5.4.0
openai>=0.27.0